outsouring of base class preprocessor

+++ dataset/extract_cuhk03.py
@@ -16,105 +16,10 @@ import shutil
 from glob import glob
 from zipfile import ZipFile
 
-
-def write_json(obj, fpath):
-    mkdir_if_missing(osp.dirname(fpath))
-    with open(fpath, 'w') as f:
-        json.dump(obj, f, indent=4, separators=(',', ': '))
-
-
-def mkdir_if_missing(dir_path):
-    try:
-        os.makedirs(dir_path)
-    except OSError as e:
-        if e.errno != errno.EEXIST:
-            raise
-
-def _pluck(identities, indices, relabel=False):
-    ret = []
-    for index, pid in enumerate(indices):
-        pid_images = identities[pid]
-        for camid, cam_images in enumerate(pid_images):
-            for fname in cam_images:
-                name = osp.splitext(fname)[0]
-                x, y, _ = map(int, name.split('_'))
-                assert pid == x and camid == y
-                if relabel:
-                    ret.append((fname, index, camid))
-                else:
-                    ret.append((fname, pid, camid))
-    return ret
-
-def read_json(fpath):
-    with open(fpath, 'r') as f:
-        obj = json.load(f)
-    return obj
-
-class DataPreprocessor(object):
-    def __init__(self, root, split_id=0):
-        self.root = root
-        self.split_id = split_id
-        self.meta = None
-        self.split = None
-        self.train, self.val, self.trainval = [], [], []
-        self.query, self.gallery = [], []
-        self.num_train_ids, self.num_val_ids, self.num_trainval_ids = 0, 0, 0
-
-    @property
-    def images_dir(self):
-        return osp.join(self.root, 'images')
-
-    def load(self, num_train=0.3, verbose=True):
-        splits = read_json(osp.join(self.root, 'splits.json'))
-        if self.split_id >= len(splits):
-            raise ValueError("split_id exceeds total splits {}"
-                             .format(len(splits)))
-        self.split = splits[self.split_id]
-        # Randomly split train / val
-        trainval_pids = np.asarray(self.split['trainval'])
-        np.random.shuffle(trainval_pids)
-        num = len(trainval_pids)
-        if isinstance(num_train, float):
-            num_train = int(round(num * num_train))
-        if num_train >= num or num_train < 0:
-            raise ValueError("num_train exceeds total identities {}"
-                             .format(num))
-
-        train_pids = sorted(trainval_pids[:num_train])
-        val_pids = sorted(trainval_pids[num_train:])
-        self.split['train'] = train_pids
-        self.split['val'] = val_pids
-
-        self.meta = read_json(osp.join(self.root, 'meta.json'))
-        identities = self.meta['identities']
-        self.train = _pluck(identities, train_pids, relabel=False)
-        self.val = _pluck(identities, val_pids, relabel=False)
-        self.trainval = _pluck(identities, trainval_pids, relabel=False)
-        self.query = _pluck(identities, self.split['query'])
-        self.gallery = _pluck(identities, self.split['gallery'])
-        self.num_train_ids = len(train_pids)
-        self.num_val_ids = len(val_pids)
-        self.num_trainval_ids = len(trainval_pids)
-
-        if verbose:
-            print(self.__class__.__name__, "dataset loaded")
-            print("  subset   | # ids | # images")
-            print("  ---------------------------")
-            print("  train    | {:5d} | {:8d}"
-                  .format(self.num_train_ids, len(self.train)))
-            print("  val      | {:5d} | {:8d}"
-                  .format(self.num_val_ids, len(self.val)))
-            print("  trainval | {:5d} | {:8d}"
-                  .format(self.num_trainval_ids, len(self.trainval)))
-            print("  query    | {:5d} | {:8d}"
-                  .format(len(self.split['query']), len(self.query)))
-            print("  gallery  | {:5d} | {:8d}"
-                  .format(len(self.split['gallery']), len(self.gallery)))
-
-    def _check_integrity(self):
-        return osp.isdir(osp.join(self.root, 'images')) and \
-               osp.isfile(osp.join(self.root, 'meta.json')) and \
-               osp.isfile(osp.join(self.root, 'splits.json'))
+from .preprocessor import write_json
+from .preprocessor import mkdir_if_missing
+from .preprocessor import read_json
+from.preprocessor import DataPreprocessor
 
 
 class CUHK03(DataPreprocessor):
@@ -123,6 +28,7 @@ class CUHK03(DataPreprocessor):
 
     def __init__(self, root, split_id=0, num_train=100, download=True):
         super(CUHK03, self).__init__(root, split_id=split_id)
+        self.exdir = None
 
         if download:
             self.download()
@@ -132,7 +38,8 @@ class CUHK03(DataPreprocessor):
                                "You can use download=True to download it.")
 
         self.load(num_train)
-        shutil.rmtree(self.exdir)
+        if self.exdir is not None:
+            shutil.rmtree(self.exdir)
 
     def deref(self, ref):
         return self.matdata[ref][:].T
